Refactor ProblemListTable and improve user registration modal
The main changes in this commit include the refactor of the ProblemListTable to use more granular components and improved the user registration modal. Additionally, minor changes were made to the project database schema and update of various dependencies in the package-lock.json file. This refinement enhances the readability of the code and improves the user experience during registration.